python/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
idna==2.8
iso-639==0.4.5
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
joblib==0.14.1
langdetect==1.0.7
lxml==4.4.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.18.1
Pillow==7.0.0
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
requests-file==1.4.3
scikit-learn==0.22.2
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
virtualenv==16.7.9
Werkzeug==0.16.0
